Refactor RF analysis: separate fitting and analysis modules
