+++ cottage_analysis/analysis/spheres/rf_fitting.py
@@ -1,5 +1,6 @@
 from functools import partial
 import gc
+
 import numpy as np
 import pandas as pd
 from scipy.optimize import curve_fit
